requirements.txt
web3
httpx
python-dotenv
asyncio
termcolor
pyotp
tweety
requests
csv